[EDIFIKANA] #383 Addressing PR Feedback in models

+++ edifikana/shared/src/commonMain/kotlin/com/cramsan/edifikana/lib/model/RentConfigModel.kt
@@ -10,7 +10,7 @@ data class RentConfigModel(
     val id: RentConfigId,
     val unitId: UnitId,
     val orgId: OrganizationId,
-    val monthlyAmount: Double,
+    val monthlyAmount: Long,
     val dueDay: Int,
     val currency: String,
     val updatedAt: Long,
